Adding support for BlowBot
Refactoring Kiiroo device classes to reduce implementation overlap
